Fix issues with studying
